gs-sidebar: deduplicate code using frontend.js
